requirements.txt
PyGithub==2.2.0
gitpython==3.1.42